file structure centered in instance folder
