+++ src/dispatch/static/dispatch/src/entity/store.js
@@ -44,27 +44,6 @@ const state = {
     },
     loading: false,
   },
-  signalEntityTable: {
-    rows: {
-      items: [],
-      total: null,
-    },
-    options: {
-      filters: {
-        created_at: {
-          start: null,
-          end: null,
-        },
-        // signal: [],
-      },
-      q: "",
-      page: 1,
-      itemsPerPage: 25,
-      sortBy: ["created_at"],
-      descending: [true],
-    },
-    loading: false,
-  },
 }
 
 const getters = {
@@ -82,59 +61,6 @@ const actions = {
       })
       .catch(() => {
         commit("SET_TABLE_LOADING", false)
-      })
-  }, 500),
-  // todo(amats): what's the difference between this and above?
-  // todo(amats): can probably put back in signal directory store with new api hookups
-  getAllEntities: debounce(({ commit, state }) => {
-    commit("SET_SIGNAL_ENTITY_TABLE_LOADING", "primary")
-    let params = SearchUtils.createParametersFromTableOptions(
-      {
-        ...state.signalEntityTable.options,
-      },
-      "Entity"
-    )
-    return EntityApi.getAll(params)
-      .then(async (response) => {
-        // Fetch signal stats for each entity in parallel
-        const fetchSignalStats = async (entity) => {
-          try {
-            const url = `/signals/stats?entity_type_id=${entity.entity_type.id}&entity_value="${entity.value}"&num_days=1000`
-            return await API.get(url)
-          } catch (error) {
-            console.error(`Error fetching signal stats for entity ${entity.name}:`, error)
-            return null
-          }
-        }
-
-        // Use Promise.all for parallel execution
-        const statsPromises = response.data.items.map(fetchSignalStats)
-        const statsResults = await Promise.all(statsPromises)
-
-        // Append signal stats to each entity item so they can be accessed in the Vue file
-        // todo(amats) can this be added to the async function for performance?
-        statsResults.forEach((result, index) => {
-          if (result) {
-            const entity = response.data.items[index]
-            // separate the stats to avoid duplicate column references when rendering
-            let instanceStats = {
-              num_signal_instances_alerted: result.data.num_signal_instances_alerted,
-              num_signal_instances_snoozed: result.data.num_signal_instances_snoozed,
-            }
-            let snoozeStats = {
-              num_snoozes_active: result.data.num_snoozes_active,
-              num_snoozes_expired: result.data.num_snoozes_expired,
-            }
-            entity.instanceStats = instanceStats
-            entity.snoozeStats = snoozeStats
-          }
-        })
-
-        commit("SET_SIGNAL_ENTITY_TABLE_LOADING", false)
-        commit("SET_SIGNAL_ENTITY_TABLE_ROWS", response.data)
-      })
-      .catch(() => {
-        commit("SET_SIGNAL_ENTITY_TABLE_LOADING", false)
       })
   }, 500),
   createEditShow({ commit }, entity) {
@@ -225,12 +151,6 @@ const mutations = {
   },
   SET_TABLE_ROWS(state, value) {
     state.table.rows = value
-  },
-  SET_SIGNAL_ENTITY_TABLE_LOADING(state, value) {
-    state.signalEntityTable.loading = value
-  },
-  SET_SIGNAL_ENTITY_TABLE_ROWS(state, value) {
-    state.signalEntityTable.rows = value
   },
   SET_DIALOG_CASE_VIEW(state, value) {
     state.dialogs.showCaseView = value
