+++ src/dispatch/static/dispatch/src/signal/filter/store.js
@@ -40,28 +40,7 @@ const state = {
       descending: [true],
     },
     loading: false,
-  },
-  snoozeTable: {
-    rows: {
-      items: [],
-      total: null,
-    },
-    options: {
-      filters: {
-        created_at: {
-          start: null,
-          end: null,
-        },
-        action: ["snooze"], // exclude deduplications
-      },
-      q: "",
-      page: 1,
-      itemsPerPage: 25,
-      sortBy: ["created_at"],
-      descending: [true],
-    },
-    loading: false,
-  },
+  }
 }
 
 const getters = {
@@ -73,18 +52,6 @@ const getters = {
 }
 
 const actions = {
-  getAllSnoozes: debounce(({ commit, state }) => {
-    commit("SET_SNOOZE_TABLE_LOADING", "primary")
-    let params = SearchUtils.createParametersFromTableOptions({ ...state.snoozeTable.options })
-    return SignalFilterApi.getAll(params)
-      .then((response) => {
-        commit("SET_SNOOZE_TABLE_LOADING", false)
-        commit("SET_SNOOZE_TABLE_ROWS", response.data)
-      })
-      .catch(() => {
-        commit("SET_SNOOZE_TABLE_LOADING", false)
-      })
-  }, 500),
   save({ commit, state }) {
     commit("SET_SELECTED_LOADING", true)
     if (!state.selected.id) {
@@ -160,12 +127,6 @@ const mutations = {
   },
   SET_DIALOG_CREATE_EDIT(state, value) {
     state.dialogs.showCreateEdit = value
-  },
-  SET_SNOOZE_TABLE_LOADING(state, value) {
-    state.snoozeTable.loading = value
-  },
-  SET_SNOOZE_TABLE_ROWS(state, value) {
-    state.snoozeTable.rows = value
   },
 
   RESET_SELECTED(state) {
