+++ src/dispatch/static/dispatch/src/signal/instance/TableInstance.vue
@@ -40,16 +40,16 @@
       </v-col>
     </v-row>
     <table-instance-triggers v-if="activeView === 'triggers'" />
-    <table-signal-entities v-if="activeView === 'entities'" />
-    <table-filter-snoozes v-if="activeView === 'snoozes'" />
+    <table-instance-entities v-if="activeView === 'entities'" />
+    <table-instance-snoozes v-if="activeView === 'snoozes'" />
   </v-container>
 </template>
 
 <script>
 import TableFilterDialog from "@/signal/TableFilterDialog.vue"
 import TableInstanceTriggers from "@/signal/instance/TableInstanceTriggers.vue"
-import TableSignalEntities from "@/entity/TableSignalEntities.vue"
-import TableFilterSnoozes from "@/signal/filter/TableFilterSnoozes.vue"
+import TableInstanceEntities from "@/signal/instance/TableInstanceEntities.vue"
+import TableInstanceSnoozes from "@/signal/instance/TableInstanceSnoozes.vue"
 
 export default {
   name: "SignalInstanceTable",
@@ -57,8 +57,8 @@ export default {
   components: {
     TableFilterDialog,
     TableInstanceTriggers,
-    TableSignalEntities,
-    TableFilterSnoozes,
+    TableInstanceEntities,
+    TableInstanceSnoozes,
   },
 
   data() {
