+++ src/dispatch/static/dispatch/src/signal/store.js
@@ -5,6 +5,8 @@ import SearchUtils from "@/search/utils"
 import SignalApi from "@/signal/api"
 import SignalFilterApi from "@/signal/filter/api"
 import EntityApi from "@/entity/api"
+// todo(amats): not like this
+import API from "@/api"
 
 const getDefaultSelectedState = () => {
   return {
@@ -95,6 +97,48 @@ const state = {
     },
     loading: false,
   },
+  signalEntityTable: {
+    rows: {
+      items: [],
+      total: null,
+    },
+    options: {
+      filters: {
+        created_at: {
+          start: null,
+          end: null,
+        },
+        // signal: [],
+      },
+      q: "",
+      page: 1,
+      itemsPerPage: 25,
+      sortBy: ["created_at"],
+      descending: [true],
+    },
+    loading: false,
+  },
+  snoozeTable: {
+    rows: {
+      items: [],
+      total: null,
+    },
+    options: {
+      filters: {
+        created_at: {
+          start: null,
+          end: null,
+        },
+        action: ["snooze"], // exclude deduplications
+      },
+      q: "",
+      page: 1,
+      itemsPerPage: 25,
+      sortBy: ["created_at"],
+      descending: [true],
+    },
+    loading: false,
+  },
 }
 
 const getters = {
@@ -131,6 +175,69 @@ const actions = {
       })
       .catch(() => {
         commit("SET_INSTANCE_TABLE_LOADING", false)
+      })
+  }, 500),
+  getAllSnoozes: debounce(({ commit, state }) => {
+    commit("SET_SNOOZE_TABLE_LOADING", "primary")
+    let params = SearchUtils.createParametersFromTableOptions({ ...state.snoozeTable.options })
+    return SignalFilterApi.getAll(params)
+      .then((response) => {
+        commit("SET_SNOOZE_TABLE_LOADING", false)
+        commit("SET_SNOOZE_TABLE_ROWS", response.data)
+      })
+      .catch(() => {
+        commit("SET_SNOOZE_TABLE_LOADING", false)
+      })
+  }, 500),
+  getAllEntities: debounce(({ commit, state }) => {
+    commit("SET_SIGNAL_ENTITY_TABLE_LOADING", "primary")
+    let params = SearchUtils.createParametersFromTableOptions(
+      {
+        ...state.signalEntityTable.options,
+      },
+      "Entity"
+    )
+    return EntityApi.getAll(params)
+      .then(async (response) => {
+        // Fetch signal stats for each entity in parallel
+        const fetchSignalStats = async (entity) => {
+          try {
+            const url = `/signals/stats?entity_type_id=${entity.entity_type.id}&entity_value="${entity.value}"&num_days=1000`
+            return await API.get(url)
+          } catch (error) {
+            console.error(`Error fetching signal stats for entity ${entity.name}:`, error)
+            return null
+          }
+        }
+
+        // Use Promise.all for parallel execution
+        const statsPromises = response.data.items.map(fetchSignalStats)
+        const statsResults = await Promise.all(statsPromises)
+
+        // Append signal stats to each entity item so they can be accessed in the Vue file
+        // todo(amats) can this be added to the async function for performance?
+        statsResults.forEach((result, index) => {
+          if (result) {
+            const entity = response.data.items[index]
+            // separate the stats to avoid duplicate column references when rendering
+            let instanceStats = {
+              num_signal_instances_alerted: result.data.num_signal_instances_alerted,
+              num_signal_instances_snoozed: result.data.num_signal_instances_snoozed,
+            }
+            let snoozeStats = {
+              num_snoozes_active: result.data.num_snoozes_active,
+              num_snoozes_expired: result.data.num_snoozes_expired,
+            }
+            entity.instanceStats = instanceStats
+            entity.snoozeStats = snoozeStats
+          }
+        })
+
+        commit("SET_SIGNAL_ENTITY_TABLE_LOADING", false)
+        commit("SET_SIGNAL_ENTITY_TABLE_ROWS", response.data)
+      })
+      .catch(() => {
+        commit("SET_SIGNAL_ENTITY_TABLE_LOADING", false)
       })
   }, 500),
   get({ commit, state }) {
@@ -232,6 +339,18 @@ const mutations = {
   },
   SET_INSTANCE_TABLE_ROWS(state, value) {
     state.instanceTable.rows = value
+  },
+  SET_SNOOZE_TABLE_LOADING(state, value) {
+    state.snoozeTable.loading = value
+  },
+  SET_SNOOZE_TABLE_ROWS(state, value) {
+    state.snoozeTable.rows = value
+  },
+  SET_SIGNAL_ENTITY_TABLE_LOADING(state, value) {
+    state.signalEntityTable.loading = value
+  },
+  SET_SIGNAL_ENTITY_TABLE_ROWS(state, value) {
+    state.signalEntityTable.rows = value
   },
   SET_DIALOG_CREATE_EDIT(state, value) {
     state.dialogs.showCreateEdit = value
